portal/requirements/base.txt
Django==4.2.11
django-environ==0.11.2
psycopg[binary]==3.1.18
gunicorn==21.2.0
django-redis==5.4.0
httpx==0.27.0
